+++ crates/router/src/routes/app.rs
@@ -9,7 +9,7 @@ use common_enums::TransactionType;
 use common_utils::crypto::Blake3;
 #[cfg(feature = "email")]
 use external_services::email::{ses::AwsSes, EmailService};
-use external_services::file_storage::FileStorageInterface;
+use external_services::{file_storage::FileStorageInterface, grpc_client::GrpcClients};
 use hyperswitch_interfaces::{
     encryption_interface::EncryptionManagementInterface,
     secrets_interface::secret_state::{RawSecret, SecuredSecret},
@@ -105,6 +105,7 @@ pub struct SessionState {
     pub tenant: Tenant,
     #[cfg(feature = "olap")]
     pub opensearch_client: Arc<OpenSearchClient>,
+    pub grpc_client: Arc<GrpcClients>,
 }
 impl scheduler::SchedulerSessionState for SessionState {
     fn get_db(&self) -> Box<dyn SchedulerInterface> {
@@ -202,6 +203,7 @@ pub struct AppState {
     pub request_id: Option<RequestId>,
     pub file_storage_client: Arc<dyn FileStorageInterface>,
     pub encryption_client: Arc<dyn EncryptionManagementInterface>,
+    pub grpc_client: Arc<GrpcClients>,
 }
 impl scheduler::SchedulerAppState for AppState {
     fn get_tenants(&self) -> Vec<String> {
@@ -351,6 +353,8 @@ impl AppState {
 
             let file_storage_client = conf.file_storage.get_file_storage_client().await;
 
+            let grpc_client = conf.grpc_client.get_grpc_client_interface().await;
+
             Self {
                 flow_name: String::from("default"),
                 stores,
@@ -367,6 +371,7 @@ impl AppState {
                 request_id: None,
                 file_storage_client,
                 encryption_client,
+                grpc_client,
             }
         })
         .await
@@ -447,6 +452,7 @@ impl AppState {
             email_client: Arc::clone(&self.email_client),
             #[cfg(feature = "olap")]
             opensearch_client: Arc::clone(&self.opensearch_client),
+            grpc_client: Arc::clone(&self.grpc_client),
         })
     }
 }
@@ -1064,6 +1070,24 @@ impl Payouts {
     }
 }
 
+#[cfg(all(feature = "oltp", feature = "v2", feature = "payment_methods_v2",))]
+impl PaymentMethods {
+    pub fn server(state: AppState) -> Scope {
+        let mut route = web::scope("/v2/payment_methods").app_data(web::Data::new(state));
+        route = route
+            .service(web::resource("").route(web::post().to(create_payment_method_api)))
+            .service(
+                web::resource("/create-intent")
+                    .route(web::post().to(create_payment_method_intent_api)),
+            )
+            .service(
+                web::resource("/{id}/confirm-intent")
+                    .route(web::post().to(confirm_payment_method_intent_api)),
+            );
+
+        route
+    }
+}
 pub struct PaymentMethods;
 
 #[cfg(all(
@@ -1190,9 +1214,16 @@ impl Organization {
             .app_data(web::Data::new(state))
             .service(web::resource("").route(web::post().to(admin::organization_create)))
             .service(
-                web::resource("/{id}")
-                    .route(web::get().to(admin::organization_retrieve))
-                    .route(web::put().to(admin::organization_update)),
+                web::scope("/{id}")
+                    .service(
+                        web::resource("")
+                            .route(web::get().to(admin::organization_retrieve))
+                            .route(web::put().to(admin::organization_update)),
+                    )
+                    .service(
+                        web::resource("/merchant_accounts")
+                            .route(web::get().to(admin::merchant_account_list)),
+                    ),
             )
     }
 }
@@ -1202,13 +1233,20 @@ pub struct MerchantAccount;
 #[cfg(all(feature = "v2", feature = "olap"))]
 impl MerchantAccount {
     pub fn server(state: AppState) -> Scope {
-        web::scope("/v2/accounts")
+        web::scope("/v2/merchant_accounts")
             .app_data(web::Data::new(state))
             .service(web::resource("").route(web::post().to(admin::merchant_account_create)))
             .service(
-                web::resource("/{id}")
-                    .route(web::get().to(admin::retrieve_merchant_account))
-                    .route(web::put().to(admin::update_merchant_account)),
+                web::scope("/{id}")
+                    .service(
+                        web::resource("")
+                            .route(web::get().to(admin::retrieve_merchant_account))
+                            .route(web::put().to(admin::update_merchant_account)),
+                    )
+                    .service(
+                        web::resource("/profiles")
+                            .route(web::get().to(admin::business_profiles_list)),
+                    ),
             )
     }
 }
@@ -1282,7 +1320,7 @@ impl MerchantConnectorAccount {
                 .service(
                     web::resource("/{merchant_id}/connectors")
                         .route(web::post().to(connector_create))
-                        .route(web::get().to(payment_connector_list)),
+                        .route(web::get().to(connector_list)),
                 )
                 .service(
                     web::resource("/{merchant_id}/connectors/{merchant_connector_id}")
@@ -1464,6 +1502,13 @@ impl Disputes {
                     .route(web::post().to(disputes::accept_dispute)),
             )
             .service(
+                web::resource("/aggregate").route(web::get().to(disputes::get_disputes_aggregate)),
+            )
+            .service(
+                web::resource("/profile/aggregate")
+                    .route(web::get().to(disputes::get_disputes_aggregate_profile)),
+            )
+            .service(
                 web::resource("/evidence")
                     .route(web::post().to(disputes::submit_dispute_evidence))
                     .route(web::put().to(disputes::attach_dispute_evidence))
@@ -1562,13 +1607,17 @@ impl BusinessProfile {
     pub fn server(state: AppState) -> Scope {
         web::scope("/v2/profiles")
             .app_data(web::Data::new(state))
-            .service(web::resource("").route(web::post().to(super::admin::business_profile_create)))
+            .service(web::resource("").route(web::post().to(admin::business_profile_create)))
             .service(
                 web::scope("/{profile_id}")
                     .service(
                         web::resource("")
-                            .route(web::get().to(super::admin::business_profile_retrieve))
-                            .route(web::put().to(super::admin::business_profile_update)),
+                            .route(web::get().to(admin::business_profile_retrieve))
+                            .route(web::put().to(admin::business_profile_update)),
+                    )
+                    .service(
+                        web::resource("/connector_accounts")
+                            .route(web::get().to(admin::connector_list)),
                     )
                     .service(
                         web::resource("/fallback_routing")
@@ -1673,8 +1722,7 @@ impl BusinessProfileNew {
                     .route(web::get().to(admin::business_profiles_list_at_profile_level)),
             )
             .service(
-                web::resource("/connectors")
-                    .route(web::get().to(admin::payment_connector_list_profile)),
+                web::resource("/connectors").route(web::get().to(admin::connector_list_profile)),
             )
     }
 }